+++ tests/testthat/test-mng01.R
@@ -4,6 +4,8 @@ test_that("mng01 works as expected with default argument values", {
   pre_data <- mng01_pre(syn_data, dataset = "adlb")
   res <- mng01_main(pre_data, dataset = "adlb")
   expect_list(res, len = 3, types = "ggplot")
+
+  skip_on_ci()
   vdiffr::expect_doppelganger("mng01 with default value", res[[1]])
 })
 
@@ -22,7 +24,6 @@ test_that("mng01 works as expected with custom argument values", {
     line_col = col
   )
   expect_list(res, len = 3, types = "ggplot")
-  vdiffr::expect_doppelganger("mng01 with custom theme", res[[1]], )
 
   bad_col <- c(
     "B: Placebo" = "black",
@@ -34,6 +35,9 @@ test_that("mng01 works as expected with custom argument values", {
     "Missing color matching for C: Combination",
     fixed = TRUE
   )
+
+  skip_on_ci()
+  vdiffr::expect_doppelganger("mng01 with custom theme", res[[1]], )
 })
 
 
@@ -49,17 +53,21 @@ test_that("mng01 works as expected with custom color set", {
   proc_data <- syn_data
   res <- run(mng01, proc_data, dataset = "adlb", line_col = col)
   expect_list(res, len = 3, types = "ggplot")
-  vdiffr::expect_doppelganger("run mng01 with custom color set", res[[1]])
 
-  res <- run(mng01, proc_data, dataset = "adlb", line_col = unname(col))
-  expect_list(res, len = 3, types = "ggplot")
-  vdiffr::expect_doppelganger("run mng01 with custom unnamed color set", res[[1]])
+  res2 <- run(mng01, proc_data, dataset = "adlb", line_col = unname(col))
+  expect_list(res2, len = 3, types = "ggplot")
+
+  skip_on_ci()
+  vdiffr::expect_doppelganger("run mng01 with custom color set", res[[1]])
+  vdiffr::expect_doppelganger("run mng01 with custom unnamed color set", res2[[1]])
 })
 
 test_that("mng01 works with table = NULL", {
   proc_data <- syn_data
   res <- run(mng01, proc_data, dataset = "adlb", table = NULL)
   expect_list(res, len = 3, types = "ggplot")
+
+  skip_on_ci()
   vdiffr::expect_doppelganger("run mng01 without table", res[[1]])
 })
 
@@ -72,6 +80,8 @@ test_that("mng01 works with combination of x variables", {
     x_var = c("AVISIT", "AVISITN")
   )
   expect_list(res, len = 3, types = "ggplot")
+
+  skip_on_ci()
   vdiffr::expect_doppelganger("run mng01 with combination of x variables", res[[1]])
 })
 
@@ -86,6 +96,8 @@ test_that("mng01 works with numeric x variable", {
     x_var = c("AVISITN")
   )
   expect_list(res, len = 3, types = "ggplot")
+
+  skip_on_ci()
   vdiffr::expect_doppelganger("run mng01 with numeric x variables", res[[1]])
 })
 
@@ -98,5 +110,7 @@ test_that("mng01 works with numeric jitter", {
     jitter = 0.7
   )
   expect_list(res, len = 3, types = "ggplot")
+
+  skip_on_ci()
   vdiffr::expect_doppelganger("run mng01 with jitter", res[[1]])
 })